feat: moved flamebearerToTree function to separate file
